+++ src/components/Atoms/Image/Icon/Icon.scss
