+++ src/components/Atoms/Image/Icon/Icon.stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, select, text } from '@storybook/addon-knobs';
 import Icon from './Icon.tsx';
-import { EIcon } from '../../../utility/enumeration';
+import { EIcon } from '../../../../utility/enumeration';
 
 const {
 	cheveronDown,
