+++ src/components/Atoms/Image/Icon/Icon.test.js
