+++ src/components/Atoms/Image/Icon/index.ts
